PO-8998, 8999, 8810 DW changes to support these reports - fixed user.groovy
git-svn-id: https://ncisvn.nci.nih.gov/svn/coppa/trunk/code@18516 553898ca-32d5-487a-af8b-6cf4ca579ab4

+++ pa/data-warehouse/groovy/user.groovy
@@ -33,7 +33,7 @@ def sql = """select ru.affiliated_org_id,
                     ru.user_last_updated_id
 		from csm_user cu
 		left outer join registry_user as ru on ru.csm_user_id = cu.user_id 
-		left outer join organization as org on org.assigned_identifier=ru.affiliated_org_id
+		left outer join organization as org on org.assigned_identifier::INTEGER=ru.affiliated_org_id
                 """
 
 def sourceConnectionPa = Sql.newInstance(properties['datawarehouse.pa.source.jdbc.url'], properties['datawarehouse.pa.source.db.username'],
